+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/JibContainerBuilder.java
@@ -23,8 +23,7 @@ import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.configuration.ContainerConfiguration;
 import com.google.cloud.tools.jib.configuration.ImageConfiguration;
 import com.google.cloud.tools.jib.configuration.LayerConfiguration;
-import com.google.cloud.tools.jib.event.DefaultEventDispatcher;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.cloud.tools.jib.image.ImageFormat;
 import com.google.cloud.tools.jib.image.LayerEntry;
@@ -475,11 +474,11 @@ public class JibContainerBuilder {
 
     BuildConfiguration buildConfiguration = toBuildConfiguration(containerizer, executorService);
 
-    EventDispatcher eventDispatcher = buildConfiguration.getEventDispatcher();
-    logSources(eventDispatcher);
+    EventHandlers eventHandlers = buildConfiguration.getEventHandlers();
+    logSources(eventHandlers);
 
     try (TimerEventDispatcher ignored =
-        new TimerEventDispatcher(eventDispatcher, containerizer.getDescription())) {
+        new TimerEventDispatcher(eventHandlers, containerizer.getDescription())) {
       BuildResult result = containerizer.createStepsRunner(buildConfiguration).run();
       return new JibContainer(result.getImageDigest(), result.getImageId());
 
@@ -523,30 +522,25 @@ public class JibContainerBuilder {
         .setToolName(containerizer.getToolName())
         .setExecutorService(executorService);
 
-    containerizer
-        .getEventHandlers()
-        .ifPresent(
-            eventHandlers ->
-                buildConfigurationBuilder.setEventDispatcher(
-                    new DefaultEventDispatcher(eventHandlers)));
+    containerizer.getEventHandlers().ifPresent(buildConfigurationBuilder::setEventHandlers);
 
     return buildConfigurationBuilder.build();
   }
 
-  private void logSources(EventDispatcher eventDispatcher) {
+  private void logSources(EventHandlers eventHandlers) {
     // Logs the different source files used.
-    eventDispatcher.dispatch(LogEvent.info("Containerizing application with the following files:"));
+    eventHandlers.dispatch(LogEvent.info("Containerizing application with the following files:"));
 
     for (LayerConfiguration layerConfiguration : layerConfigurations) {
       if (layerConfiguration.getLayerEntries().isEmpty()) {
         continue;
       }
 
-      eventDispatcher.dispatch(
+      eventHandlers.dispatch(
           LogEvent.info("\t" + capitalizeFirstLetter(layerConfiguration.getName()) + ":"));
 
       for (LayerEntry layerEntry : layerConfiguration.getLayerEntries()) {
-        eventDispatcher.dispatch(LogEvent.info("\t\t" + layerEntry.getSourceFile()));
+        eventHandlers.dispatch(LogEvent.info("\t\t" + layerEntry.getSourceFile()));
       }
     }
   }