+++ jib-core/src/main/java/com/google/cloud/tools/jib/configuration/BuildConfiguration.java
@@ -17,7 +17,7 @@
 package com.google.cloud.tools.jib.configuration;
 
 import com.google.cloud.tools.jib.cache.Cache;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
 import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
@@ -61,10 +61,7 @@ public class BuildConfiguration {
     private ImmutableList<LayerConfiguration> layerConfigurations = ImmutableList.of();
     private Class<? extends BuildableManifestTemplate> targetFormat = DEFAULT_TARGET_FORMAT;
     private String toolName = DEFAULT_TOOL_NAME;
-    private EventDispatcher eventDispatcher =
-        jibEvent -> {
-          /* No-op EventDispatcher. */
-        };
+    private EventHandlers eventHandlers = new EventHandlers();
     @Nullable private ExecutorService executorService;
 
     private Builder() {}
@@ -192,13 +189,13 @@ public class BuildConfiguration {
     }
 
     /**
-     * Sets the {@link EventDispatcher} to dispatch events with.
+     * Sets the {@link EventHandlers} to dispatch events with.
      *
-     * @param eventDispatcher the {@link EventDispatcher}
+     * @param eventHandlers the {@link EventHandlers}
      * @return this
      */
-    public Builder setEventDispatcher(EventDispatcher eventDispatcher) {
-      this.eventDispatcher = eventDispatcher;
+    public Builder setEventHandlers(EventHandlers eventHandlers) {
+      this.eventHandlers = eventHandlers;
       return this;
     }
 
@@ -242,7 +239,7 @@ public class BuildConfiguration {
       switch (missingFields.size()) {
         case 0: // No errors
           if (Preconditions.checkNotNull(baseImageConfiguration).getImage().usesDefaultTag()) {
-            eventDispatcher.dispatch(
+            eventHandlers.dispatch(
                 LogEvent.warn(
                     "Base image '"
                         + baseImageConfiguration.getImage()
@@ -261,7 +258,7 @@ public class BuildConfiguration {
               offline,
               layerConfigurations,
               toolName,
-              eventDispatcher,
+              eventHandlers,
               Preconditions.checkNotNull(executorService));
 
         case 1:
@@ -314,7 +311,7 @@ public class BuildConfiguration {
   private final boolean offline;
   private final ImmutableList<LayerConfiguration> layerConfigurations;
   private final String toolName;
-  private final EventDispatcher eventDispatcher;
+  private final EventHandlers eventHandlers;
   private final ExecutorService executorService;
 
   /** Instantiate with {@link #builder}. */
@@ -330,7 +327,7 @@ public class BuildConfiguration {
       boolean offline,
       ImmutableList<LayerConfiguration> layerConfigurations,
       String toolName,
-      EventDispatcher eventDispatcher,
+      EventHandlers eventHandlers,
       ExecutorService executorService) {
     this.baseImageConfiguration = baseImageConfiguration;
     this.targetImageConfiguration = targetImageConfiguration;
@@ -343,7 +340,7 @@ public class BuildConfiguration {
     this.offline = offline;
     this.layerConfigurations = layerConfigurations;
     this.toolName = toolName;
-    this.eventDispatcher = eventDispatcher;
+    this.eventHandlers = eventHandlers;
     this.executorService = executorService;
   }
 
@@ -376,8 +373,8 @@ public class BuildConfiguration {
     return toolName;
   }
 
-  public EventDispatcher getEventDispatcher() {
-    return eventDispatcher;
+  public EventHandlers getEventHandlers() {
+    return eventHandlers;
   }
 
   public ExecutorService getExecutorService() {
@@ -451,7 +448,7 @@ public class BuildConfiguration {
 
   private RegistryClient.Factory newRegistryClientFactory(ImageConfiguration imageConfiguration) {
     return RegistryClient.factory(
-            getEventDispatcher(),
+            getEventHandlers(),
             imageConfiguration.getImageRegistry(),
             imageConfiguration.getImageRepository())
         .setAllowInsecureRegistries(getAllowInsecureRegistries())