+++ jib-core/src/main/java/com/google/cloud/tools/jib/frontend/CredentialRetrieverFactory.java
@@ -19,7 +19,7 @@ package com.google.cloud.tools.jib.frontend;
 import com.google.cloud.tools.jib.api.ImageReference;
 import com.google.cloud.tools.jib.configuration.credentials.Credential;
 import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.JibEvent;
 import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.cloud.tools.jib.registry.credentials.CredentialHelperNotFoundException;
@@ -57,13 +57,13 @@ public class CredentialRetrieverFactory {
    * Creates a new {@link CredentialRetrieverFactory} for an image.
    *
    * @param imageReference the image the credential are for
-   * @param eventDispatcher an event dispatcher for dispatching log events
+   * @param eventHandlers an event handlers for dispatching log events
    * @return a new {@link CredentialRetrieverFactory}
    */
   public static CredentialRetrieverFactory forImage(
-      ImageReference imageReference, EventDispatcher eventDispatcher) {
+      ImageReference imageReference, EventHandlers eventHandlers) {
     return new CredentialRetrieverFactory(
-        imageReference, eventDispatcher, DockerCredentialHelper::new);
+        imageReference, eventHandlers, DockerCredentialHelper::new);
   }
 
   /**
@@ -74,20 +74,20 @@ public class CredentialRetrieverFactory {
    */
   public static CredentialRetrieverFactory forImage(ImageReference imageReference) {
     return new CredentialRetrieverFactory(
-        imageReference, ignored -> {}, DockerCredentialHelper::new);
+        imageReference, new EventHandlers(), DockerCredentialHelper::new);
   }
 
-  private final EventDispatcher eventDispatcher;
+  private final EventHandlers eventHandlers;
   private final ImageReference imageReference;
   private final DockerCredentialHelperFactory dockerCredentialHelperFactory;
 
   @VisibleForTesting
   CredentialRetrieverFactory(
       ImageReference imageReference,
-      EventDispatcher eventDispatcher,
+      EventHandlers eventHandlers,
       DockerCredentialHelperFactory dockerCredentialHelperFactory) {
     this.imageReference = imageReference;
-    this.eventDispatcher = eventDispatcher;
+    this.eventHandlers = eventHandlers;
     this.dockerCredentialHelperFactory = dockerCredentialHelperFactory;
   }
 
@@ -221,7 +221,7 @@ public class CredentialRetrieverFactory {
     return () -> {
       try {
         Optional<Credential> dockerConfigCredentials =
-            dockerConfigCredentialRetriever.retrieve(eventDispatcher);
+            dockerConfigCredentialRetriever.retrieve(eventHandlers);
         if (dockerConfigCredentials.isPresent()) {
           dispatchEvent(
               LogEvent.info(
@@ -253,9 +253,6 @@ public class CredentialRetrieverFactory {
   }
 
   private void dispatchEvent(JibEvent jibEvent) {
-    if (eventDispatcher == null) {
-      return;
-    }
-    eventDispatcher.dispatch(jibEvent);
+    eventHandlers.dispatch(jibEvent);
   }
 }