+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/ManifestPusher.java
@@ -20,7 +20,7 @@ import com.google.api.client.http.HttpMethods;
 import com.google.api.client.http.HttpResponseException;
 import com.google.cloud.tools.jib.api.DescriptorDigest;
 import com.google.cloud.tools.jib.blob.Blobs;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.cloud.tools.jib.hash.Digests;
 import com.google.cloud.tools.jib.http.BlobHttpContent;
@@ -66,17 +66,17 @@ class ManifestPusher implements RegistryEndpointProvider<DescriptorDigest> {
   private final RegistryEndpointRequestProperties registryEndpointRequestProperties;
   private final BuildableManifestTemplate manifestTemplate;
   private final String imageTag;
-  private final EventDispatcher eventDispatcher;
+  private final EventHandlers eventHandlers;
 
   ManifestPusher(
       RegistryEndpointRequestProperties registryEndpointRequestProperties,
       BuildableManifestTemplate manifestTemplate,
       String imageTag,
-      EventDispatcher eventDispatcher) {
+      EventHandlers eventHandlers) {
     this.registryEndpointRequestProperties = registryEndpointRequestProperties;
     this.manifestTemplate = manifestTemplate;
     this.imageTag = imageTag;
-    this.eventDispatcher = eventDispatcher;
+    this.eventHandlers = eventHandlers;
   }
 
   @Override
@@ -143,7 +143,7 @@ class ManifestPusher implements RegistryEndpointProvider<DescriptorDigest> {
     }
 
     // The received digest is not as expected. Warns about this.
-    eventDispatcher.dispatch(
+    eventHandlers.dispatch(
         LogEvent.warn(makeUnexpectedImageDigestWarning(expectedDigest, receivedDigests)));
     return expectedDigest;
   }